+++ CHANGELOG.md
@@ -11,7 +11,8 @@ Version 1.3.x
 - Added support for multi-dimensional quantile queries
 - Allow to quickly check a benchmark from the [Quantitative Verification Benchmark Set](http://qcomp.org/benchmarks/) using the --qvbs option.
 - Added script resources/examples/download_qvbs.sh to download the QVBS.
-- If an option is unknown, Storm now prints a hint to similar option names.
+- If an option is unknown, Storm now suggests similar option names.
+- Flagged several options as 'advanced' to clean up the `--help`-message. Use `--help all` to display a complete list of options.
 - Support for the new `round` operator in the PRISM language
 - Support for parsing of exact time bounds for properties, e.g., `P=? [F=27 "goal"]`
 - JANI: Allow bounded types for constants